pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "Cython>=3.0.11"]
build-backend = "setuptools.build_meta"

[project]
name = "image_cloud"
version = "2025.0.3"
authors = [
    { name = "Larry Page", email = "lawrence.page@live.com" },
]
description = "Creation of a collage of images scaled based on relative weights."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"
license = { text = "MIT License" }
dependencies = [ 
    "pillow>=11.1.0",
    "numpy>=2.2.1",
    "matplotlib>=3.10.0",
    "webcolors>=24.11.1"
]

[project.urls]
Homepage = "https://github.com/lpage30/image_cloud"

[project.scripts]
generate_imagecloud = "imagecloud_clis.generate_cli:generate"
layout_imagecloud = "imagecloud_clis.layout_cli:layout"

[tool.setuptools.packages.find]
where = ["src"]